feat(data): added a lot of data for reductions and desynth
