+++ apps/client/src/app/modules/list/data/extractor/desynths-extractor.ts
@@ -3,10 +3,12 @@ import { ItemData } from '../../../../model/garland-tools/item-data';
 import { DataType } from '../data-type';
 import { Item } from '../../../../model/garland-tools/item';
 import { GarlandToolsService } from '../../../../core/api/garland-tools.service';
+import { LazyDataService } from '../../../../core/data/lazy-data.service';
+import { uniq } from 'lodash';
 
 export class DesynthsExtractor extends AbstractExtractor<number[]> {
 
-  constructor(gt: GarlandToolsService) {
+  constructor(gt: GarlandToolsService, private lazyData: LazyDataService) {
     super(gt);
   }
 
@@ -19,11 +21,11 @@ export class DesynthsExtractor extends AbstractExtractor<number[]> {
   }
 
   protected canExtract(item: Item): boolean {
-    return item.desynthedFrom !== undefined && item.desynthedFrom.length > 0;
+    return (item.desynthedFrom !== undefined && item.desynthedFrom.length > 0) || this.lazyData.data.desynth[item.id] !== undefined;
   }
 
   protected doExtract(item: Item, itemData: ItemData): number[] {
-    return item.desynthedFrom;
+    return uniq([...(item.desynthedFrom || []), ...(this.lazyData.data.desynth[item.id] || [])]);
   }
 
 }
