+++ apps/client/src/app/modules/list/data/extractor/reduced-from-extractor.ts
@@ -3,10 +3,12 @@ import { ItemData } from '../../../../model/garland-tools/item-data';
 import { DataType } from '../data-type';
 import { Item } from '../../../../model/garland-tools/item';
 import { GarlandToolsService } from '../../../../core/api/garland-tools.service';
+import { LazyDataService } from '../../../../core/data/lazy-data.service';
+import { uniq } from 'lodash';
 
 export class ReducedFromExtractor extends AbstractExtractor<number[]> {
 
-  constructor(gt: GarlandToolsService) {
+  constructor(gt: GarlandToolsService, private lazyData: LazyDataService) {
     super(gt);
   }
 
@@ -19,11 +21,11 @@ export class ReducedFromExtractor extends AbstractExtractor<number[]> {
   }
 
   protected canExtract(item: Item): boolean {
-    return item.reducedFrom !== undefined && item.reducedFrom.length > 0;
+    return (item.reducedFrom !== undefined && item.reducedFrom.length > 0) || this.lazyData.data.reduction[item.id] !== undefined;
   }
 
   protected doExtract(item: Item, itemData: ItemData): any[] {
-    return item.reducedFrom.map(reduction => itemData.getPartial(reduction.toString(), 'item'));
+    return uniq([...(item.reducedFrom || []), ...(this.lazyData.data.reduction[item.id] || [])]);
   }
 
 }
