+++ data-extraction/src/input/shb-fish-desynth.csv
@@ -1,0 +1,143 @@
+Item,r1,r2,r3,r4,r5
+Abyssal Snail,Fine Sand,Water Crystal,Purple Pigment,,
+Albino Caiman,Fine Sand,Water Crystal,Grey Pigment,Nightworld Bronze Piece,
+Albino Garfish,Fine Sand,Water Crystal,,,
+Amber Lamprey,Fine Sand,Water Crystal,Red Pigment,,
+Ankle Snipper,Fine Sand,Water Crystal,,,
+Anpa's Handmaid,Fine Sand,Water Crystal,,,
+Aquabloom,Fine Sand,Water Crystal,,,
+Asteroidea,Fine Sand,Water Crystal,,,
+Azure Sea Spider,Fine Sand,Water Crystal,Blue Pigment,,
+Bedskipper,Fine Sand,Water Crystal,,,
+Big-eye,,Water Crystal,Blue Pigment,Nightworld Bronze Piece,
+Blood Cloud,Fine Sand,Water Crystal,Yellow Pigment,Nightworld Bronze Piece,
+Blood-eyed Frog,Fine Sand,Water Crystal,Green Pigment,Nightworld Bronze Piece,
+Blue Lightning,Fine Sand,Water Crystal,,,
+Bonefish,Fine Sand,Water Crystal,Grey Pigment,,
+Bubble Angler,Fine Sand,Water Crystal,Grey Pigment,,
+Caliban,Fine Sand,Water Crystal,Blue Pigment,,
+Cerulean Loach,Fine Sand,Water Crystal,Blue Pigment,Nightworld Bronze Piece,
+Cherry Herring,Fine Sand,Water Crystal,,,
+Clawbow,Fine Sand,Water Crystal,Red Pigment,,
+Clean Saucer,Fine Sand,Water Crystal,Grey Pigment,,
+Clown Tetra,Fine Sand,Water Crystal,,,
+Crimson Sea Spider,Fine Sand,Water Crystal,Red Pigment,,
+Crystal Knife,Fine Sand,Water Crystal,Blue Pigment,,
+Crystarium Tetra,,,,,
+Cyan Sea Devil,Fine Sand,Water Crystal,,,
+Daisy Turban,Fine Sand,Water Crystal,Yellow Pigment,,
+Dandyfish,,Water Crystal,Purple Pigment,,
+Darkroot,Fine Sand,Water Crystal,,,
+Deep Purple Coral,CANNOT DESYNTH,,,,
+Desert Dustfish,Fine Sand,Water Crystal,Yellow Pigment,,
+Desert Saw,Fine Sand,Water Crystal,,,
+Diamond Pipira,,Water Crystal,,Nightworld Bronze Piece,
+Diamondtongue,Fine Sand,Water Crystal,,,
+Dohn Horn,Fine Sand,Water Crystal,Yellow Pigment,Nightworld Bronze Piece,
+Elven Spear,Fine Sand,Water Crystal,Grey Pigment,,
+Eulmore Butterfly,Fine Sand,Water Crystal,Red Pigment,,
+Everdark Bass,,Water Crystal,Green Pigment,,
+Finned Eggplant,Fine Sand,Water Crystal,,,
+Flowering Kelpie,Fine Sand,Water Crystal,,,
+Fuathgobbler,Fine Sand,Water Crystal,Red Pigment,Nightworld Bronze Piece,
+Garik Crab,Fine Sand,Water Crystal,Grey Pigment,,
+Gatorl's Bead,Fine Sand,Water Crystal,,,
+Gazing Glass,Fine Sand,Water Crystal,,,
+Geayi,Fine Sand,Water Crystal,,,
+Ghoulfish,Fine Sand,Water Crystal,Purple Pigment,,
+Gilded Batfish,Fine Sand,Water Crystal,,,Grade 7 Dark Matter
+Glass Eel,Fine Sand,Water Crystal,,,
+Golden Lobster,Fine Sand,Water Crystal,,,
+Gourmand Crab,Fine Sand,Water Crystal,Blue Pigment,,
+Gravel Mussel,Fine Sand,Water Crystal,,,
+Grey Carp,Fine Sand,Water Crystal,Grey Pigment,,
+Grey Skipper,Fine Sand,Water Crystal,Grey Pigment,Nightworld Bronze Piece,
+Hard Candy,Fine Sand,Water Crystal,Blue Pigment,,
+Hermit Crab,,,,,
+Hermit's Hood,Fine Sand,Water Crystal,Brown Pigment,,
+Hoodwinker,Fine Sand,Water Crystal,,,
+Hornhelm,Fine Sand,Water Crystal,Green Pigment,,
+Hucho Taimen,Fine Sand,Water Crystal,,,
+Hunter's Arrow,Fine Sand,Water Crystal,,,
+Jacketed Snail,Fine Sand,Water Crystal,,,
+Jenanna's Tear,Fine Sand,Water Crystal,Blue Pigment,Nightworld Bronze Piece,
+Jester Fish,,Water Crystal,Red Pigment,,
+Kholusian Flounder,Fine Sand,Water Crystal,Brown Pigment,Nightworld Bronze Piece,
+Kholusian Wrasse,Fine Sand,Water Crystal,,,
+Lakeland Cod,Fine Sand,Water Crystal,,Nightworld Bronze Piece,
+Lakelouse,,,,,
+Lakethistle,Fine Sand,Water Crystal,Purple Pigment,,
+Laxan Carp,Fine Sand,Water Crystal,Grey Pigment,,
+Lemonfish,Fine Sand,Water Crystal,,,
+Lilac Goby,Fine Sand,Water Crystal,Purple Pigment,Nightworld Bronze Piece,
+Lozatl Pirarucu,Fine Sand,Water Crystal,,,
+Maidenhair,Fine Sand,Water Crystal,,,
+Measan Deala,Fine Sand,Water Crystal,Yellow Pigment,,Paralyzing Potion
+Megapiranha,Fine Sand,Water Crystal,,,
+Milky Coral,CANNOT DESYNTH,,,,
+Minstrelfish,,Water Crystal,Brown Pigment,,
+Mirrorfish,Fine Sand,Water Crystal,Red Pigment,,Blood Red Dye
+Misteye,Fine Sand,Water Crystal,,,
+Misty Killifish,Fine Sand,Water Crystal,Grey Pigment,,
+Mock Pixie,Fine Sand,Water Crystal,Purple Pigment,,
+Nabaath Manta,Fine Sand,Water Crystal,Brown Pigment,,
+Night's Bass,Fine Sand,Water Crystal,,,
+Noblefish,,Water Crystal,Purple Pigment,,
+Oathfish,Fine Sand,Water Crystal,Red Pigment,,
+Official Ball,Fine Sand,Water Crystal,,,
+Ondo Harpoon,Fine Sand,Water Crystal,,,
+Ondobane,,Water Crystal,Blue Pigment,,
+Queensgown,,Water Crystal,,,Glamour Prism
+Paradise Crab,,Water Crystal,Grey Pigment,,
+Petalfish,Fine Sand,Water Crystal,Purple Pigment,,
+Pixie Fish,Fine Sand,Water Crystal,Red Pigment,,
+Platinum Bream,,Water Crystal,Grey Pigment,,
+Purple Ghost,,,,,
+Rabbit Skipper,Fine Sand,Water Crystal,Grey Pigment,,
+Rainbow Shrimp,Fine Sand,Water Crystal,Green Pigment,,
+Rak'tika Trout,Fine Sand,Water Crystal,,Nightworld Bronze Piece,
+Razorfish,Fine Sand,Water Crystal,,,
+Rebel,Fine Sand,Water Crystal,,,
+Red Hammerhead,Fine Sand,Water Crystal,Red Pigment,,Bluespirit File
+Robber Crab,,Water Crystal,Red Pigment,,
+Ronkan Pleco,Fine Sand,Water Crystal,Yellow Pigment,,
+Saint Fathric's Ire,Fine Sand,Water Crystal,Red Pigment,,
+Sand Egg,Fine Sand,Water Crystal,,,
+Sauldia Ruby,Fine Sand,Water Crystal,Purple Pigment,,
+Seatrap,,Water Crystal,Brown Pigment,,
+Sepia Sole,,Water Crystal,,Nightworld Bronze Piece,
+Shade Axolotl,Fine Sand,Water Crystal,Green Pigment,Nightworld Bronze Piece,Purple Pigment
+Shade Gudgeon,Fine Sand,Water Crystal,,,
+Shadow Crab,Fine Sand,Water Crystal,,Nightworld Bronze Piece,
+Shapeshifter,,Water Crystal,,Nightworld Bronze Piece,
+Shellfracture Kelp,Fine Sand,Water Crystal,Yellow Pigment,Nightworld Bronze Piece,
+Silver Kitten,Fine Sand,Water Crystal,,,Grade 7 Dark Matter
+Sinspitter,Fine Sand,Water Crystal,,,
+Skeletonfish,Fine Sand,Water Crystal,,,
+Skulleater,Fine Sand,Water Crystal,,,
+Skykisser,,Water Crystal,Grey Pigment,,
+Source Octopus,Fine Sand,Water Crystal,,,
+South Kholusian Cod,Fine Sand,Water Crystal,Brown Pigment,Nightworld Bronze Piece,
+Spotted Blue-eye,,Water Crystal,,Nightworld Bronze Piece,
+Spotted Ctenopoma,Fine Sand,Water Crystal,,,
+Stargazer,,Water Crystal,Grey Pigment,,
+Stippled Eel,Fine Sand,Water Crystal,,,
+Stormfish,Fine Sand,Water Crystal,Blue Pigment,Nightworld Bronze Piece,
+Sweetflesh Oyster,Fine Sand,Water Crystal,,,
+Sycorax,Fine Sand,Water Crystal,,,
+Thorned Lizard,Fine Sand,Water Crystal,,,
+Toffee Snail,Fine Sand,Water Crystal,Yellow Pigment,,
+Treescale,Fine Sand,Water Crystal,,,
+Viola Clam,,Water Crystal,Purple Pigment,,
+Vicejaw,Fine Sand,Water Crystal,,,
+Voeburt Salamander,,Water Crystal,Green Pigment,,
+Wandering Catfish,,,,,
+Wardenfish,Fine Sand,Water Crystal,Purple Pigment,,Quickarm Materia V
+Water Ball,Fine Sand,Water Crystal,,,
+Watts Trout,,Water Crystal,Brown Pigment,,
+Weedy Seadragon,,Water Crystal,Yellow Pigment,,
+Wild Red Betta,,Water Crystal,Red Pigment,,
+Wimple Carp,Fine Sand,Water Crystal,,,
+Winged Hatchetfish,Fine Sand,Water Crystal,Grey Pigment,,
+Xanthic Bass,Fine Sand,Water Crystal,Yellow Pigment,,
+Yellow Pipira,Fine Sand,Water Crystal,Yellow Pigment,Nightworld Bronze Piece,Potent Paralyzing Potion
