--- data-extraction/xivapi/.gitignore
@@ -1,5 +1,0 @@
-# Created by .ignore support plugin (hsz.mobi)
-output
-input/memory_data.csv
-input/lgb/*
-progress.json
