--- data-extraction/xivapi/input/monster-drops.csv
@@ -1,17 +1,0 @@
-Monster,Drops
-Smilodon,Smilodon Skin
-Hoptrap,Hoptrap leaf
-Silkmoth,Silkmoth Scales
-Island rail,Rail Tenderloin
-Island Wolf,Island Wolf Fang
-Ovim Billy,"Ovim Meat, Ovim Fleece"
-Ovim Nanny,"Ovim Meat, Ovim Fleece"
-Flame Zonure,Zonure Skin
-Rainbow Lorikeet,"Lorikeet Egg, Lorikeet Down"
-Green Glider,Green Glider Skin
-Vampire Vine,Vampire Vine Sap
-Vampire Cup,Vampire Cup Vine
-Atrociraptor,Atrociraptor Skin
-Tempest Swallow,Sea Swallow Skin
-Hydrozoan,Hydrozoan Umbrella
-Cubus,Cubus Flesh
