--- data-extraction/xivapi/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "xivapi-mapper",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js",
-    "generate-sitemap": "node seo/sitemap-generator.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "csv-parser": "^1.12.0",
-    "request": "^2.88.0",
-    "rxjs": "^6.3.3",
-    "cli-progress": "3.6.0"
-  }
-}
